Implement functionality to calc req'd and then sel nearest match in menu
So far only implemented on Shutter menu, i.e. have to select ISO and
Aperture, it will then use the equations to calculate the exact required
shutter speed, but then select the nearest value in the menu.

In order to do that added these two functions:

- nearest_match
- fraction_to_double

The latter needed, because need to convert "1/1000", etc into float. For
Aperture menu items can just strip off the "f/".

Seems to work for Sunny 16.