Try separate win and menu functions
instead of trying to wrap everything in one. This way I know I can
return the correct thing and I know I'm setting things as expected.

As opposed to previous attempt where I was (trying) to use variable
pointers, but with pointers to pointers, etc. And was obviously a bit
beyond me.